refactor: remove storage state view aggregate
