+++ polylogue/pipeline/services/rendering.py
@@ -12,7 +12,7 @@ from typing import TYPE_CHECKING
 from polylogue.lib.metrics import read_current_rss_mb
 from polylogue.logging import get_logger
 from polylogue.protocols import OutputRenderer, ProgressCallback
-from polylogue.storage.state_views import RenderFailurePayload
+from polylogue.storage.run_state import RenderFailurePayload
 
 if TYPE_CHECKING:
     from polylogue.storage.backends.async_sqlite import SQLiteBackend
