+++ polylogue/sources/decoder_zip.py
@@ -10,7 +10,7 @@ from typing import BinaryIO, cast
 from polylogue.lib.artifact_taxonomy import classify_artifact_path
 from polylogue.logging import get_logger
 from polylogue.storage.blob_store import get_blob_store
-from polylogue.storage.state_views import CursorStatePayload
+from polylogue.storage.cursor_state import CursorStatePayload
 from polylogue.types import Provider
 
 from .cursor import _record_cursor_failure
