+++ polylogue/storage/backends/queries/raw_state.py
@@ -8,7 +8,7 @@ from typing import Any
 import aiosqlite
 
 from polylogue.storage.backends.connection import _build_source_scope_filter
-from polylogue.storage.state_views import UNSET, RawConversationStateUpdate, _RawStateUnset
+from polylogue.storage.raw_state_models import UNSET, RawConversationStateUpdate, _RawStateUnset
 from polylogue.types import Provider, ValidationMode, ValidationStatus
 
 EFFECTIVE_RAW_PROVIDER_SQL = "COALESCE(payload_provider, provider_name)"
