+++ polylogue/storage/repository_raw.py
@@ -7,8 +7,8 @@ from typing import TYPE_CHECKING
 
 from polylogue.storage.backends.queries import artifacts as artifacts_q
 from polylogue.storage.backends.queries import raw as raw_queries
+from polylogue.storage.raw_state_models import RawConversationState, RawConversationStateUpdate
 from polylogue.storage.repository_contracts import RepositoryBackendProtocol
-from polylogue.storage.state_views import RawConversationState, RawConversationStateUpdate
 from polylogue.storage.store import (
     ArtifactObservationRecord,
     RawConversationRecord,
