--- polylogue/storage/state_views.py
@@ -1,56 +1,0 @@
-"""Temporary aggregate re-export for split storage runtime view modules."""
-
-from __future__ import annotations
-
-from polylogue.storage.archive_views import (
-    ConversationRenderProjection,
-    ExistingConversation,
-)
-from polylogue.storage.artifact_views import ArtifactCohortSummary
-from polylogue.storage.cursor_state import CursorFailurePayload, CursorStatePayload
-from polylogue.storage.raw_state_models import (
-    UNSET,
-    RawConversationState,
-    RawConversationStateUpdate,
-    _RawStateUnset,
-)
-from polylogue.storage.run_state import (
-    DriftBucket,
-    DriftBucketPayload,
-    PlanCounts,
-    PlanCountsPayload,
-    PlanDetails,
-    PlanDetailsPayload,
-    PlanResult,
-    RenderFailurePayload,
-    RunCounts,
-    RunCountsPayload,
-    RunDrift,
-    RunDriftPayload,
-    RunResult,
-)
-
-__all__ = [
-    "ArtifactCohortSummary",
-    "ConversationRenderProjection",
-    "CursorFailurePayload",
-    "CursorStatePayload",
-    "DriftBucket",
-    "DriftBucketPayload",
-    "ExistingConversation",
-    "PlanCounts",
-    "PlanCountsPayload",
-    "PlanDetails",
-    "PlanDetailsPayload",
-    "PlanResult",
-    "RawConversationState",
-    "RawConversationStateUpdate",
-    "RenderFailurePayload",
-    "RunCounts",
-    "RunCountsPayload",
-    "RunDrift",
-    "RunDriftPayload",
-    "RunResult",
-    "UNSET",
-    "_RawStateUnset",
-]
